improve mode unit tests
